# swaggerSchoolTemplate.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "School Management App"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/loginschool": {
      "x-swagger-router-controller": "schoolController",
      "post": {
        "tags": [
          "School"
        ],
        "operationId": "loginSchoolAdmin",
        "summary": "Get the scholl details",
        "description": "Get the school details",
        "parameters": [
          {
            "name": "schoolData",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/GetSchoolResponse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School Added successfully",
            "schema": {
              "$ref": "#/definitions/GetSchoolResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "GetSchoolResponse": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}